# use a buffer pool & don't allocate ints