@kevinburke 2016-06-21
This would break compatibility for go1.1 and go1.2