@inconshreveable 2016-06-21
actually, i'm okay only supporting Go1.3+. let's just change the travis.yml file to reflect that